Use UUIDV4 to generate random usernames instead of randomIntBetween
More and more often, the repository's CI tests would fail on a test
user creation. This failure is caused by the pre-existence of the
username used during the creation. By the looks of it, the collisions
happen because we have already created a lot of test users over time.

Unfortunately the test api doesn't offer the possibility to delete
a user, yet. Thus, we switch the random number generation to UUIDv4
which has, hopefully, enough cardinality to not have this problem
anymore in the upcoming future, and until we implement the ability
to delete users in the test API.