Support for alternative bundle file name
As of today it was only possible to have the jslib.k6.io page point to
a package's bundle file named `index.js`. With the introduction of the
aws package, this became an issue as its main bundle is called `aws.js`
and it also exposes a variety of others.

This commit allows to use the `bundle-filename` property inside the
`supported.json` file to define the name of the bundle file.